rui_zhu/requirements_Jenny.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
astroid==2.11.7
async-generator==1.10
attrs==21.4.0
autograd==1.4
autograd-gamma==0.5.0
backcall==0.2.0
bleach==4.1.0
boto3==1.23.10
botocore==1.26.10
cffi==1.15.1
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
dtw-python==1.2.2
entrypoints==0.4
fastdtw==0.3.4
formulaic==0.2.4
future==0.18.2
importlib-metadata==4.8.3
importlib-resources==5.4.0
interface-meta==1.2.5
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.17.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
lifelines==0.26.4
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.7.0
mistune==0.8.4
mpmath==1.2.1
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numba==0.53.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2==2.9.3
ptyprocess==0.7.0
pyampd==0.0.1
pycparser==2.21
Pygments==2.12.0
pylint==2.13.9
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
pytimeparse==1.1.8
pytz==2022.1
pyzmq==23.2.0
roerich==0.2
ruptures==1.1.7
s3transfer==0.5.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
snfpy==0.2.2
sns==0.1
sympy==1.9
terminado==0.12.1
testpath==0.6.0
threadpoolctl==3.1.0
tomli==1.2.3
torch==1.10.2
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
typed-ast==1.5.4
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
xgboost==1.5.2
zipp==3.6.0
